fix: update id api docs
