--- src/id/resources/profile/attributes/appearance.ts
@@ -1,22 +1,0 @@
-import { ProfileAppearance } from '../../../../types'
-import Resource from '../../resource'
-
-export default class ProfileAttributesAppearance extends Resource {
-  get (): Promise<ProfileAppearance> {
-    return this.client.call({
-      method: 'get',
-      endpoint: '/profiles/me/attributes/appearance'
-    }).then((response) => {
-      return response.data
-    })
-  }
-
-  update (): Promise<ProfileAppearance> {
-    return this.client.call({
-      method: 'put',
-      endpoint: '/profiles/me/attributes/appearance'
-    }).then((response) => {
-      return response.data
-    })
-  }
-}
