+++ src/id/resources/profile/attributes/appearance/index.ts
@@ -1,0 +1,22 @@
+import { ProfileAppearance } from '../../../../../types'
+import Resource from '../../../resource'
+
+export default class ProfileAttributesAppearance extends Resource {
+  get (): Promise<ProfileAppearance> {
+    return this.client.call({
+      method: 'get',
+      endpoint: '/profiles/me/attributes/appearance'
+    }).then((response) => {
+      return response.data
+    })
+  }
+
+  update (): Promise<ProfileAppearance> {
+    return this.client.call({
+      method: 'put',
+      endpoint: '/profiles/me/attributes/appearance'
+    }).then((response) => {
+      return response.data
+    })
+  }
+}
