+++ src/id/resources/profile/ledgers/index.ts
@@ -1,0 +1,22 @@
+import { OrderLedger, TransactionLedger } from '../../../../types'
+import Resource from '../../resource'
+
+export default class ProfileLedgers extends Resource {
+  get (ledgerID: string): Promise<(TransactionLedger | OrderLedger)> {
+    return this.client.call({
+      method: 'get',
+      endpoint: `/profiles/me/ledgers/${ledgerID}`
+    }).then((response) => {
+      return response.data
+    })
+  }
+
+  list (): Promise<(TransactionLedger | OrderLedger)[]> {
+    return this.client.call({
+      method: 'get',
+      endpoint: '/profiles/me/ledgers'
+    }).then((response) => {
+      return response.data
+    })
+  }
+}
