--- src/id/resources/profile/transaction-claims.ts
@@ -1,32 +1,0 @@
-import { TransactionClaim, TransactionClaimsResponse, RequestParams } from '../../../types'
-import Resource from '../resource'
-
-export default class ProfileTransactions extends Resource {
-  get (claimId: number): Promise<TransactionClaim> {
-    return this.client.call({
-      method: 'get',
-      endpoint: `/profiles/me/transactions/claims/${claimId}`
-    }).then((response) => {
-      return response.data
-    })
-  }
-
-  list (params?: RequestParams): Promise<TransactionClaimsResponse> {
-    return this.client.call({
-      method: 'get',
-      endpoint: '/profiles/me/transactions/claims',
-      params
-    }).then((response) => {
-      return response
-    })
-  }
-
-  delete (claimId: number): Promise<TransactionClaimsResponse> {
-    return this.client.call({
-      method: 'delete',
-      endpoint: `/profiles/me/transactions/claims/${claimId}`
-    }).then((response) => {
-      return response.data
-    })
-  }
-}
